% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{changeStyle}
\alias{changeStyle}
\title{Display 3D Strucutre Styles Of Protein In shiny App}
\usage{
changeStyle(style = "cartoon", pdbid, color = "grey")
}
\arguments{
\item{style}{protein structure styles such as cartoon, stick, sphere. cartoon shows structure in loops. sphere shows atoms in sphere. stick shows bonds as sticks. default style is cartoon.}

\item{color}{color of the structure such as blue, green, yellow, and etc. Users can also use hex color. Default color is grey.}

\item{pdbID}{protein database id such as "1A2B"}
}
\value{
a shiny app with protein structure shown.
}
\description{
open a 3dmol.js website with different molecule structure styles shown with white surface prediction.
}
\examples{

# show 1a2b in red stick protein strucutre.
changeStyle("stick", "1a2b", "red")

#show 1a2b in blue sphere protein strucutre.
changeStyle("sphere", "1a2b", "333CFF")
}
\references{
Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL
\emph{Bioinformatics} (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.

Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2020). shiny: Web Application Framework for R. \emph{R package version 1.5.0}.
https://CRAN.R-project.org/package=shiny
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
