% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{color_Pfam}
\alias{color_Pfam}
\title{Display 3D Strucutre Of Protein In shiny App}
\usage{
color_Pfam(PDBid)
}
\arguments{
\item{pdbID}{protein database id such as "1A2B"}
}
\value{
a shiny app with protein structure shown and Pfam structure annoted.
}
\description{
open a 3dmol.js contained shiny app with molecule structure shown and Pfam structure annoted.
}
\examples{

# Generating protein structure with Pfam structures in blue
color_Pfam("1a2b")

}
\references{
Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL
\emph{Bioinformatics} (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.

Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2020). shiny: Web Application Framework for R. \emph{R package version 1.5.0}.
https://CRAN.R-project.org/package=shiny.

Grant, B.J. et al. (2006) \emph{Bioinformatics} 22, 2695--2696.
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
