% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{compareTwoPDB}
\alias{compareTwoPDB}
\title{anaylsis the strucuture similarity of two pdb file}
\usage{
compareTwoPDB(pdbId1, pdbId2)
}
\arguments{
\item{pdbId1}{protein database id such as "1A2B"}

\item{pdbId2}{protein database id such as "1A2B"}
}
\value{
visualizations of the two pdb file
}
\description{
anaylsis the strucuture similarity of two pdb file
}
\examples{

# Generating protein structure 
compareTwoPDB('1a2b','1a1b')

}
\references{
1. Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL
\emph{Bioinformatics} (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
