% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{parseSpec}
\alias{parseSpec}
\title{reformate the input}
\usage{
parseSpec(style, color)
}
\arguments{
\item{style}{protein structure styles such as cartoon, stick, sphere. cartoon shows structure in loops. sphere shows atoms in sphere. stick shows bonds as sticks. default style is cartoon.}

\item{color}{color of the structure such as blue, green, yellow, and etc. Users can also use hex color. Default color is grey.}
}
\value{
new formated string
}
\description{
reformate the input
}
\examples{

# show 1a2b in red stick protein strucutre.
parseSpec("stick", "red")

#show 1a2b in blue sphere protein strucutre.
changeStyle("sphere", "1a2b", "333CFF")
}
\references{
R Core Team (2019). R: A language and environment for statistical computing. \emph{R Foundation for Statistical Computing}, Vienna, Austria. URL https://www.R-project.org/.
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
