% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPdb3D.R
\name{runPdb3D}
\alias{runPdb3D}
\title{Launch Shiny App For Package pdb3D}
\usage{
runPdb3D()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app allows user to choose a input file, the app
will calculate the difference among those genes and generate
a plot shows the distribution for all differences.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
pdb3D::runPdb3D()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL *Bioinformatics* (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.
}
\author{
Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}
}
