% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{show3Dmol}
\alias{show3Dmol}
\title{Display 3D Strucutre Of Protein In shiny App}
\usage{
show3Dmol(pdbID)
}
\arguments{
\item{pdbID}{protein database id such as "1A2B"}
}
\value{
a 3dmol object in htmlwidgets app with protein structure shown.
}
\description{
open a 3dmol.js contained shiny app with molecule structure shown
}
\examples{

# Generating protein structure 
show3Dmol("1a2b")

}
\references{
1. Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL
\emph{Bioinformatics} (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.
2. Wei Su (2020). r3dmol: Create Interactive 3D Visualizations of Molecular Data. R
package version 0.1.0. https://github.com/swsoyee/r3dmol
3. Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
