% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_3D.R
\name{show_3Dmol}
\alias{show_3Dmol}
\title{Display 3D Strucutre Of Protein In shiny App}
\usage{
show_3Dmol(PDBid)
}
\arguments{
\item{pdbID}{protein database id such as "1A2B"}
}
\value{
a shiny app with protein structure shown.
}
\description{
open a 3dmol.js contained shiny app with molecule structure shown
}
\examples{

# Generating protein structure 
show_3Dmol("1a2b")

}
\references{
Nicholas Rego and David Koes 3Dmol.js: molecular visualization with WebGL
\emph{Bioinformatics} (2015) 31 (8): 1322-1324 doi:10.1093/bioinformatics/btu829.

Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2020). shiny: Web Application Framework for R. \emph{R package version 1.5.0}.
https://CRAN.R-project.org/package=shiny
}
\author{
{Wenkai Cao, \email{Wenkai.cao@mail.utoronto.ca}}
}
